completed part 4

+++ part4/blog-backend/controllers/blogs.js
@@ -1,25 +1,56 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user')
     response.json(blogs.map(blog => blog.toJSON()))
 
   })
+
+blogsRouter.get('/:id', async (request, response) => {
+    const blog = await Blog.findById(request.params.id).populate('user')
+    response.json(blog.toJSON())
+  })  
   
 blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
-    result = await blog.save();
-    response.status(201).json(result)
+    
+  let blog = request.body
 
+  const user = request.user
+  if (!user){
+    return response.status(401).json({
+      error: 'Not a valid user'
+    })
+  }
+  blog = new Blog({...blog, user: user.id})
+
+  savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
   })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
 
-  })
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  if (!user){
+    return response.status(401).json({
+      error: 'Not a valid user'
+    })
+  }
+  
+  if ( blog.user.toString() === user.id.toString() ){
+    user.blogs = user.blogs.filter(blogid => blogid.toString() !== request.params.id)
+      await Blog.findByIdAndDelete(request.params.id)
+      await user.save()
+      response.status(204).end()
+    }   
+})
 
 blogsRouter.put('/:id', async (request, response) => {
